// apps/react-frontend/src/api/client/models/processing-method.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ProcessingMethod
 */
export interface ProcessingMethod {
    /**
     * 
     * @type {string}
     * @memberof ProcessingMethod
     */
    'code': string;
    /**
     * 
     * @type {string}
     * @memberof ProcessingMethod
     */
    'description': string;
}